+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log
 
+## [0.0.2](https://github.com/ivosights/sociomile-ios-sdk/releases/tag/0.0.2) (16/04/2024)
+
+## Fixing
+
+- Navigation Controller
+- Change example project UI
+
 ## [0.0.1](https://github.com/ivosights/sociomile-ios-sdk/releases/tag/0.0.1) (16/04/2024)
 
 ## New Features